add interop probing api client
