+++ src/main/java/it/pagopa/pn/client/b2b/pa/service/IPnDowntimeLogsClient.java
@@ -7,13 +7,10 @@ import it.pagopa.pn.client.b2b.web.generated.openapi.clients.externalDowntimeLog
 import it.pagopa.pn.client.b2b.web.generated.openapi.clients.externalDowntimeLogs.model.PnStatusUpdateEvent;
 import java.time.OffsetDateTime;
 import java.util.List;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.client.RestClientException;
 
 
 public interface IPnDowntimeLogsClient {
-
-  public ResponseEntity<Void> getEserviceStatus() throws RestClientException;
     PnStatusResponse currentStatus() throws RestClientException;
     LegalFactDownloadMetadataResponse getLegalFact(String legalFactId) throws RestClientException;
     PnStatusResponse status() throws RestClientException;
