+++ src/main/java/it/pagopa/pn/client/b2b/pa/service/impl/PnDowntimeLogsExternalClientImpl.java
@@ -2,7 +2,6 @@ package it.pagopa.pn.client.b2b.pa.service.impl;
 
 import it.pagopa.pn.client.b2b.pa.service.IPnDowntimeLogsClient;
 import it.pagopa.pn.client.b2b.web.generated.openapi.clients.externalDowntimeLogs.ApiClient;
-import it.pagopa.pn.client.b2b.web.generated.openapi.clients.externalDowntimeLogs.api.InteropProbingApi;
 import it.pagopa.pn.client.b2b.web.generated.openapi.clients.externalDowntimeLogs.api.ReadApi;
 import it.pagopa.pn.client.b2b.web.generated.openapi.clients.externalDowntimeLogs.api.WriteApi;
 import it.pagopa.pn.client.b2b.web.generated.openapi.clients.externalDowntimeLogs.model.LegalFactDownloadMetadataResponse;
@@ -13,7 +12,6 @@ import it.pagopa.pn.client.b2b.web.generated.openapi.clients.externalDowntimeLog
 import java.time.OffsetDateTime;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestTemplate;
@@ -23,7 +21,6 @@ import org.springframework.web.client.RestTemplate;
 public class PnDowntimeLogsExternalClientImpl implements IPnDowntimeLogsClient {
     private final ReadApi readApi;
     private final WriteApi writeApi;
-    private final InteropProbingApi interopProbingApi;
 
 
     public PnDowntimeLogsExternalClientImpl(RestTemplate restTemplate,
@@ -32,12 +29,6 @@ public class PnDowntimeLogsExternalClientImpl implements IPnDowntimeLogsClient {
                                             @Value("${pn.webapi.external.user-agent}")String userAgent) {
         this.readApi = new ReadApi( newApiClient( restTemplate, basePath, bearerToken,userAgent) );
         this.writeApi = new WriteApi( newApiClient( restTemplate, basePath, bearerToken,userAgent) );
-        this.interopProbingApi = new InteropProbingApi(
-            newApiClient(restTemplate, basePath, bearerToken, userAgent));
-    }
-
-    public ResponseEntity<Void> getEserviceStatus() throws RestClientException {
-        return this.interopProbingApi.getEserviceStatusWithHttpInfo();
     }
 
     private static ApiClient newApiClient(RestTemplate restTemplate, String basePath, String bearerToken, String userAgent ) {
