+++ src/test/java/it/pagopa/pn/cucumber/steps/pa/InteropProbingSteps.java
@@ -2,7 +2,7 @@ package it.pagopa.pn.cucumber.steps.pa;
 
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
-import it.pagopa.pn.client.b2b.pa.service.IPnDowntimeLogsClient;
+import it.pagopa.pn.client.b2b.pa.service.IPnInteropProbingClient;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Assertions;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,17 +11,17 @@ import org.springframework.http.ResponseEntity;
 @Slf4j
 public class InteropProbingSteps {
 
-  private final IPnDowntimeLogsClient downtimeLogsClient;
+  private final IPnInteropProbingClient IPnInteropProbingClient;
   private ResponseEntity<Void> probingResponse;
 
   @Autowired
-  public InteropProbingSteps(IPnDowntimeLogsClient downtimeLogsClient) {
-    this.downtimeLogsClient = downtimeLogsClient;
+  public InteropProbingSteps(IPnInteropProbingClient IPnInteropProbingClient) {
+    this.IPnInteropProbingClient = IPnInteropProbingClient;
   }
 
   @When("viene chiamato il servizio di probing per interop")
   public void probingService() {
-    probingResponse = downtimeLogsClient.getEserviceStatus();
+    probingResponse = IPnInteropProbingClient.getEserviceStatus();
   }
 
   @Then("la chiamata al servizio di probing per interop restituisce {int}")
